+++ cpp/src/sdk/grpc/ClientProxy.cpp
@@ -83,6 +83,7 @@ ClientProxy::CreateTable(const TableSchema &param) {
         ::milvus::grpc::TableSchema schema;
         schema.mutable_table_name()->set_table_name(param.table_name);
         schema.set_dimension(param.dimension);
+        schema.set_index_file_size(param.index_file_size);
 
         return client_ptr_->CreateTable(schema);
     } catch (std::exception &ex) {
@@ -119,7 +120,6 @@ ClientProxy::CreateIndex(const IndexParam &index_param) {
                 index_param.table_name);
         grpc_index_param.mutable_index()->set_index_type((int32_t)index_param.index_type);
         grpc_index_param.mutable_index()->set_nlist(index_param.nlist);
-        grpc_index_param.mutable_index()->set_index_file_size(index_param.index_file_size);
         grpc_index_param.mutable_index()->set_metric_type(index_param.metric_type);
         return client_ptr_->CreateIndex(grpc_index_param);
 
@@ -272,6 +272,7 @@ ClientProxy::DescribeTable(const std::string &table_name, TableSchema &table_sch
 
         table_schema.table_name = grpc_schema.table_name().table_name();
         table_schema.dimension = grpc_schema.dimension();
+        table_schema.index_file_size = grpc_schema.index_file_size();
 
         return status;
     } catch (std::exception &ex) {
@@ -362,7 +363,6 @@ ClientProxy::DescribeIndex(const std::string &table_name, IndexParam &index_para
         Status status = client_ptr_->DescribeIndex(grpc_table_name, grpc_index_param);
         index_param.index_type = (IndexType)(grpc_index_param.mutable_index()->index_type());
         index_param.nlist = grpc_index_param.mutable_index()->nlist();
-        index_param.index_file_size = grpc_index_param.mutable_index()->index_file_size();
         index_param.metric_type = grpc_index_param.mutable_index()->metric_type();
 
         return status;